+++ CHANGELOG.md
@@ -194,3 +194,4 @@
 - 2023-12-02: Features Makefile, frontier, bettercontact, descend, +2
 - 2023-12-03: Test find email, README, chat, Makefile, +2
 - 2023-12-03: Docs campaign, discover, LICENSE, mod, +4
+- 2023-12-04: Cache follow up, users, contacts, frontier, +5